feat: add ListDatastores endpoint
This endpoint can be used to show all datastores, filter by name and/or type, and include various pieces of metadata about the datastore in question. By default, minimal data from the helm release is included, but the results can be enriched with:

- env group info
- semi-structured metadata
- datastore status
- the config used to create the datastore

Note that this also allows us to deprecate the Status endpoint, as that can optionally be retrieved on this api call.